fix[MC_BIN_UTILS]: fix mc_bin_utils for multiple file generation
This fix a bug in mc_bin_utils where the code would crash when
trying to generate more than 10 file.

The bug was due to the dynamic renaming of the file changing
adding 0 to previous files when passing a power of 10.

In the case of number going higher than 10 the code was looking to
rename number 10 which was not already logged