Add VirtualTorqueSensor
Add a new `VirtualTorqueSensor` based on `Device`
This sensor provide an interface for joint torque, floating base wrench
estimation for later use in `mc_rtc` for external forces compensation or
explicit compliance.
The sensor has been added to JVRC1 by default to be used in the tests.
It is tested in `test_mc_rbdyn` with
`hasDevice<mc_rbdyn::VirtualTorqueSensor>("ExtTorquesVirtSensor")`.
The `testCannonicalRobot` has been updated to account for the extra
device.